Fixes TypEr crash when doing old files cleanup
I reported this bug some months ago. Kostis Sagonas acknowledged that
there was a bug but not action was taken. So I would like to insist
politely that a fix should be included for this silly bug.

The bug is very simple and the solution *does not alter* in any way the
core functionality of TypEr.

The thing is that when TypEr starts, it tries to clean up the old files
generated by itself, but if any of those files is not there, TypEr will
crash because of this:

	ok = file:delete(OldFile)

So, it is as simple as to replace that line with:

	file:delete(OldFile)